+++ Runtime/Integrations/Godot/Util/Hjson/Extensions/JsonObjectGodotExtensions.cs
@@ -3,12 +3,20 @@
 
 #if GODOT4_3_OR_GREATER
 
+using System;
+using System.Collections.Specialized;
 using Godot;
+using Ju.Handlers;
 
 namespace Ju.Hjson
 {
 	public static class JsonObjectGodotExtensions
 	{
+		public static void Subscribe(this JsonObject obj, Node node, Action<NotifyCollectionChangedEventArgs> action, bool alwaysActive = false)
+		{
+			obj.Subscribe(new NodeLinkHandler(node, alwaysActive), action);
+		}
+
 		public static Godot.Color GetValue(this JsonObject obj, string key, Godot.Color defaultValue = default(Godot.Color))
 		{
 			return obj.ContainsKey(key) ? obj[key].GetValue(defaultValue) : defaultValue;