+++ Runtime/Util/Hjson/IO/BaseReader.cs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
+// Copyright (c) 2021-2025 Juan Delgado (@JuDelCo)
 // Copyright (c) 2014-2016 Christian Zangl
 // Copyright (c) 2001-2003 Ximian, Inc
-// Copyright (c) 2021-2025 Juan Delgado (@JuDelCo)
 // Based on System.Json from https://github.com/mono/mono (MIT X11)
 
 using System;