+++ Runtime/Util/Hjson/JsonPrimitive.cs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
+// Copyright (c) 2021-2025 Juan Delgado (@JuDelCo)
 // Copyright (c) 2014-2016 Christian Zangl
 // Copyright (c) 2001-2003 Ximian, Inc
-// Copyright (c) 2021-2025 Juan Delgado (@JuDelCo)
 // Based on System.Json from https://github.com/mono/mono (MIT X11)
 
 using System;
@@ -73,7 +73,7 @@ namespace Ju.Hjson
 			switch (JsonType)
 			{
 				case JsonType.String:
-					return ((string)value) ?? "";
+					return ((string) value) ?? "";
 				case JsonType.Number:
 #if __MonoCS__ // mono bug ca 2014
           if (value is decimal)
@@ -85,7 +85,7 @@ namespace Ju.Hjson
           }
 #endif
 					// use ToLowerInvariant() to convert E to e
-					return ((IFormattable)value).ToString("G", NumberFormatInfo.InvariantInfo).ToLowerInvariant();
+					return ((IFormattable) value).ToString("G", NumberFormatInfo.InvariantInfo).ToLowerInvariant();
 				default:
 					throw new InvalidOperationException();
 			}